clean up formatter
